数据建模--nifi删除

+++ fk-datamodel/fk-datamodel-pojo/src/main/java/com/fisk/datamodel/vo/DataModelTableVO.java
@@ -14,7 +14,7 @@ public class DataModelTableVO {
     /**
      * 表集合id
      */
-    public List<Integer> ids;
+    public List<Long> ids;
     /**
      * 表类型：维度/事实..
      */
